Restructure NotFound folder
